style(CarouselView.scss): update styling for slick-dots and slick-slider buttons to improve visibility and consistency with light theme
The changes were made to improve the visibility and consistency of the buttons in the CarouselView component. The background color of the slick-dots buttons was updated to $color-paper and $color-coal for the light theme. Similarly, the color of the slick-slider buttons was updated to $color-paper and $color-coal for the light theme.